fixed: build error and also build the change_owner_and_threshold fn() in src/lib.rs
