+++ multisig/programs/multisig/src/instructions/cancel_tx.rs
@@ -1,5 +1,8 @@
 use anchor_lang::prelude::*;
-use crate::state::{Multisig, Transaction};
+use crate::{
+    state::{Multisig, Transaction, TransactionAccount},
+    error::ErrorCode
+};
 // use crate::constants::SEED;
 
     #[derive(Accounts)]
