+++ multisig/programs/multisig/src/instructions/init_multisig.rs
@@ -1,6 +1,8 @@
 use anchor_lang::prelude::*;
-use crate::state::{Multisig};
-
+use crate::{
+    state::{Multisig, Transaction, TransactionAccount},
+    error::ErrorCode
+};
 #[derive(Accounts)]
 pub struct InitMultisig<'info> {
     #[account(zero, signer)]
